quinx-gui/backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.3
python-jose[cryptography]==3.3.0
pydantic-settings>=2.5.0
pydantic[email]>=2.11.0
python-multipart==0.0.9
python-dotenv>=1.0.0
openpyxl>=3.1.0
